\name{elSAXdb-package}
\alias{elSAXdb-package}
\docType{package}
\title{R companion to DNA electrostatic database}
\description{Transform time series into textual and SAX representation compatible with PostgreSQL Decima datatype.
Communicate with the database, implements general CRUD procedures and functions for custom SQL queries.}
\details{
\tabular{ll}{Package: \tab elSAXdb\cr
Type: \tab Package\cr
Title: \tab R companion to DNA electrostatic database\cr
Version: \tab 0.0-2\cr
Date: \tab 2012-11-23\cr
Author: \tab Anatoly Sorokin, Evgenia Temlyakova 2012\cr
Maintainer: \tab Anatoly Sorokin <lptolik@gmail.com>\cr
License: \tab GPL (>= 2)\cr
Depends: \tab R (>= 2.14), RPostgreSQL,reldna,\cr
Sugests: \tab \cr}
}
\author{Anatoly Sorokin, Evgenia Temlyakova 2012}

\keyword{ package }


