\name{findElSeq}
\alias{findElSeq}
\title{findElSeq}
\description{main search function: it calculates the potential profile, wrap it into Decima TS string and submit query to the database }
\usage{findElSeq(seq, ref = 271)}
\arguments{
  \item{seq}{sequence to be processed}
  \item{ref}{reference point to align with promoter TSS}
}



\author{Anatoly Sorokin, Evgenia Temlyakova 2012}





