\name{getexp}
\alias{getexp}
\title{getexp}
\description{Helper function for mapply on regexp}
\usage{getexp(str, greg)}
\arguments{
  \item{str}{string to analyse}
  \item{greg}{result of grexpr invocation}
}

\value{list of matches found by regular expression in the string}

\author{Anatoly Sorokin, Evgenia Temlyakova 2012}





