\name{parseArray}
\alias{parseArray}
\title{parseArray}
\description{Converts string of numbers similar to C static arrays to vector of doubles}
\usage{parseArray(string)}
\arguments{
  \item{string}{string of numbers similar to C static array definition}
}

\value{vector of doubles}

\author{Anatoly Sorokin, Evgenia Temlyakova 2012}





