\name{shiftTSwindow}
\alias{shiftTSwindow}
\title{shiftTSwindow}
\description{Shifts time series to new window position}
\usage{shiftTSwindow(ts, offset = 0, length = -1)}
\arguments{
  \item{ts}{time series to shift}
  \item{offset}{new start position}
  \item{length}{length of new window}
}

\value{shifted time series}

\author{Anatoly Sorokin, Evgenia Temlyakova 2012}
\note{if length of new window is negative, then new window will last}




