\name{wrapSignal}
\alias{wrapSignal}
\title{wrapSignal}
\description{Funcion converts numerical vector into database TS string representation. Signas assume to be regulaly spaced.}
\usage{wrapSignal(sig)}
\arguments{
  \item{sig}{signal to convert}
}



\author{Anatoly Sorokin, Evgenia Temlyakova 2012}





